bs_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adbs_cli"
version = "1.0.0"
authors = [
  { name="Patrick Nisperos", email="pnispero@slac.stanford.edu" },
  { name="Jerry Katzung", email="katzung@slac.stanford.edu" },
]
description = "Accelerator Directorate Build System Command Line Interface"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "click==8.1.7",
  "requests==2.31.0",
  "pyyaml==6.0.1",
  "inquirer==3.2.4",
  "ansible_runner==2.4.0",
  "ansible-core==2.17.1",
  "GitPython==3.1.43"
]

[project.scripts]
bs = "adbs_cli.bs_main:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/ad-build-test/BuildSystem"
Issues = "https://github.com/ad-build-test/BuildSystem/issues"